requirements.txt
langgraph>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
jinja2>=3.1.0
rich>=13.0.0

# Optional LLM providers (install the one you need):
# pip install langchain-anthropic    # Claude
# pip install langchain-google-genai # Gemini
# pip install langchain-ollama       # Local models (Ollama)
